split out common/__init__ and add a new decorator to mark tools that should be sent over to the UI
